Sending a fcm message from an eclipse standlaone java app to Android app
At last, I was able to send a downstream fcm message using a java app
(BTW it is a gradle project) on my laptop (which will be later a server
because I want to be able to receive upstream messages).
I had to add a lot of referenced libraries (jar files) which I took a
snapshot of.