Getting location updates periodically even if the app is in background or dead, or even if mobile restarted
About every 10 minutes I get location update, rarely every minute as
well. That was made possible using the well fitting WorkManager.
It's weird though, I was expecting to get it every minute. The
'periodic' functionality of FusedLocationProviderClient is probably
diminished here, in favor of WorkManager. So maybe
FusedLocationProviderClient is not the best api to use, still it works
fine.

https://developer.android.com/guide/background

https://developer.android.com/topic/libraries/architecture/workmanager/basics

https://developer.android.com/topic/libraries/architecture/workmanager/how-to/recurring-work

P.S : SendLocationToServer class has compile errors, but it's not
important